Made JsonMng static
PlayersJsonManager and MonsterAPI can be static, since they are utility classes.

A singleton wasn't needed cuz there was not need for an instance, no state to be saved or anything.

Note:pass the managers and services around so it remains clear which classes depend on which(and it's easier to test, and no memory leaks from static references(singleton/static))